Always create `/etc/shells` if it's missing
Without an existing `/etc/shells` it's no possible to select a prefered
shell with the `Terminal: Select Default Profile` command,
a profile added with `terminal.integrated.profiles.linux` will not be
listed, and it's not possible to set it as default with
`terminal.integrated.defaultProfile.linux`.
So lets always create `/etc/shells`, even if the user blocked host
access.