Remove `zypak` module
`zypak` is included in `org.electronjs.Electron2.BaseApp` since 2020 [1] and
was already removed when the vscode flatpak started to use the electron base
app [2].

It was added back in PR #300 [3], which was not really necessary, because at
the time the PR was merged, `zypak` was already updated in the base app. The
author of the PR requested to remove `zypak` again [4], but that never
happened.

[1] https://github.com/flathub/org.electronjs.Electron2.BaseApp/pull/10
[2] https://github.com/flathub/com.visualstudio.code/commit/4538e48ac42819f7a4a1d37abb98f676512bb8a0
[3] https://github.com/flathub/com.visualstudio.code/pull/300
[4] https://github.com/flathub/com.visualstudio.code/pull/300#issuecomment-1092647335