Fix startup on Wayland systems
vscode has been updated to Electron 39, which uses Wayland by default on
systems where the environment variable `XDG_SESSION_TYPE` is set to `wayland`.
On these systems, vscode no longer starts if the Wayland socket is not present.
This commit makes the Wayland socket available in the sandbox to fix this
issue.